move file picker to model management
