+++ android/app/src/main/AndroidManifest.xml
@@ -66,7 +66,7 @@
         </activity>
 
         <activity
-            android:name=".autopilot.BackHandlingFilePickerActivity"
+            android:name=".modelManagement.BackHandlingFilePickerActivity"
             android:label="@string/app_name"
             android:theme="@style/FilePickerTheme">
             <intent-filter>
