+++ android/app/src/main/java/org/openbot/modelManagement/ModelManagementFragment.java
@@ -29,7 +29,6 @@ import java.util.Arrays;
 import java.util.List;
 import java.util.stream.Collectors;
 import org.openbot.R;
-import org.openbot.autopilot.BackHandlingFilePickerActivity;
 import org.openbot.databinding.FragmentModelManagementBinding;
 import org.openbot.main.OnItemClickListener;
 import org.openbot.tflite.Model;
